fixed api mock
